automation jobs: execute queue messages directly
