+++ apps/web/app/api/automation-jobs/execute/route.ts
@@ -1,25 +1,11 @@
-import { z } from "zod";
 import { withError } from "@/utils/middleware";
 import { withQstashOrInternal } from "@/utils/qstash";
-import prisma from "@/utils/prisma";
 import {
-  AutomationJobRunStatus,
-  MessagingProvider,
-} from "@/generated/prisma/enums";
-import { createEmailProvider } from "@/utils/email/provider";
-import { getAutomationJobMessage } from "@/utils/automation-jobs/message";
-import {
-  AutomationJobConfigurationError,
-  sendAutomationMessageToSlack,
-} from "@/utils/automation-jobs/slack";
-import { isActivePremium } from "@/utils/premium";
-import { getUserPremium } from "@/utils/user/get";
+  executeAutomationJobBody,
+  executeAutomationJobRun,
+} from "@/utils/automation-jobs/execute";
 
 export const maxDuration = 300;
-
-const executeAutomationJobBody = z.object({
-  automationJobRunId: z.string().min(1, "Automation job run ID is required"),
-});
 
 export const POST = withError(
   "automation-jobs/execute",
@@ -36,216 +22,9 @@ export const POST = withError(
       return new Response("Invalid payload", { status: 400 });
     }
 
-    const { automationJobRunId } = validation.data;
-
-    const run = await prisma.automationJobRun.findUnique({
-      where: { id: automationJobRunId },
-      include: {
-        automationJob: {
-          include: {
-            messagingChannel: {
-              include: {
-                emailAccount: {
-                  select: {
-                    id: true,
-                    userId: true,
-                    email: true,
-                    name: true,
-                    about: true,
-                    account: {
-                      select: {
-                        provider: true,
-                      },
-                    },
-                    user: {
-                      select: {
-                        aiProvider: true,
-                        aiModel: true,
-                        aiApiKey: true,
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
+    return executeAutomationJobRun({
+      automationJobRunId: validation.data.automationJobRunId,
+      logger,
     });
-
-    if (!run) {
-      logger.warn("Automation job run not found", { automationJobRunId });
-      return new Response("Automation job run not found", { status: 404 });
-    }
-
-    if (run.status === AutomationJobRunStatus.RUNNING) {
-      logger.info("Automation job run is already RUNNING", {
-        automationJobRunId,
-        createdAt: run.createdAt,
-      });
-      return new Response("Run currently running", { status: 409 });
-    } else if (run.status !== AutomationJobRunStatus.PENDING) {
-      logger.info("Automation job run already processed", {
-        automationJobRunId,
-        status: run.status,
-      });
-      return new Response("Run already processed", { status: 200 });
-    }
-
-    const claimed = await prisma.automationJobRun.updateMany({
-      where: {
-        id: automationJobRunId,
-        status: AutomationJobRunStatus.PENDING,
-      },
-      data: {
-        status: AutomationJobRunStatus.RUNNING,
-      },
-    });
-
-    if (claimed.count === 0) {
-      logger.info("Automation job run was claimed by another worker", {
-        automationJobRunId,
-      });
-      return new Response("Run already claimed", { status: 409 });
-    }
-
-    const runLogger = logger.with({
-      automationJobRunId,
-      automationJobId: run.automationJobId,
-      emailAccountId: run.automationJob.emailAccountId,
-    });
-
-    try {
-      if (!run.automationJob.enabled) {
-        runLogger.info("Skipping automation job run because job is disabled");
-        await markAutomationJobRunSkipped({
-          automationJobRunId,
-          error: "Automation job is disabled",
-        });
-
-        return new Response("Automation job disabled", { status: 200 });
-      }
-
-      const premium = await getUserPremium({
-        userId: run.automationJob.messagingChannel.emailAccount.userId,
-      });
-      if (!isActivePremium(premium)) {
-        runLogger.info(
-          "Skipping automation job run because owner is not premium",
-        );
-        await markAutomationJobRunSkipped({
-          automationJobRunId,
-          error: "Owner no longer has active premium",
-        });
-
-        return new Response("Owner no longer has active premium", {
-          status: 200,
-        });
-      }
-
-      if (!run.automationJob.messagingChannel.isConnected) {
-        runLogger.info(
-          "Skipping automation job run because messaging channel is disconnected",
-        );
-        await markAutomationJobRunSkipped({
-          automationJobRunId,
-          error: "Messaging channel is disconnected",
-        });
-
-        return new Response("Messaging channel disconnected", { status: 200 });
-      }
-
-      if (
-        run.automationJob.messagingChannel.provider !== MessagingProvider.SLACK
-      ) {
-        throw new AutomationJobConfigurationError(
-          "Unsupported messaging provider for automation job",
-        );
-      }
-
-      const provider =
-        run.automationJob.messagingChannel.emailAccount.account.provider;
-      if (!provider) {
-        throw new AutomationJobConfigurationError(
-          "Email provider is not connected",
-        );
-      }
-
-      const emailProvider = await createEmailProvider({
-        emailAccountId: run.automationJob.emailAccountId,
-        provider,
-        logger: runLogger,
-      });
-
-      const outboundMessage = await getAutomationJobMessage({
-        prompt: run.automationJob.prompt,
-        emailProvider,
-        emailAccount: run.automationJob.messagingChannel.emailAccount,
-        logger: runLogger,
-      });
-
-      const slackResult = await sendAutomationMessageToSlack({
-        channel: run.automationJob.messagingChannel,
-        text: outboundMessage,
-        logger: runLogger,
-      });
-
-      await prisma.automationJobRun.update({
-        where: { id: automationJobRunId },
-        data: {
-          status: AutomationJobRunStatus.SENT,
-          processedAt: new Date(),
-          outboundMessage,
-          providerMessageId: slackResult.messageId,
-          error: null,
-        },
-      });
-
-      return new Response("Automation job executed", { status: 200 });
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error
-          ? error.message
-          : "Failed to execute automation job";
-      const isConfigurationError =
-        error instanceof AutomationJobConfigurationError;
-
-      runLogger.error("Automation job execution failed", { error });
-
-      await prisma.automationJobRun.update({
-        where: { id: automationJobRunId },
-        data: {
-          status: isConfigurationError
-            ? AutomationJobRunStatus.SKIPPED
-            : AutomationJobRunStatus.FAILED,
-          processedAt: new Date(),
-          error: errorMessage,
-        },
-      });
-
-      return new Response(
-        isConfigurationError
-          ? "Automation job skipped due to configuration"
-          : "Automation job execution failed",
-        { status: isConfigurationError ? 200 : 500 },
-      );
-    }
   }),
 );
-
-async function markAutomationJobRunSkipped({
-  automationJobRunId,
-  error,
-}: {
-  automationJobRunId: string;
-  error: string;
-}) {
-  await prisma.automationJobRun.update({
-    where: { id: automationJobRunId },
-    data: {
-      status: AutomationJobRunStatus.SKIPPED,
-      processedAt: new Date(),
-      error,
-    },
-  });
-}
